Moved PlotLineOrBand into a subfolder.
