+++ ts/Core/Axis/PlotLineOrBand/PlotBandOptions.d.ts
@@ -17,11 +17,11 @@
 import type {
     AlignValue,
     VerticalAlignValue
-} from '../Renderer/AlignObject';
-import type ColorString from '../Color/ColorString';
-import type CSSObject from '../Renderer/CSSObject';
-import type DashStyleValue from '../Renderer/DashStyleValue';
-import type FormatUtilities from '../FormatUtilities';
+} from '../../Renderer/AlignObject';
+import type ColorString from '../../Color/ColorString';
+import type ColorType from '../../Color/ColorType';
+import type CSSObject from '../../Renderer/CSSObject';
+import type FormatUtilities from '../../FormatUtilities';
 import type PlotLineOrBand from './PlotLineOrBand';
 
 /* *
@@ -30,7 +30,7 @@ import type PlotLineOrBand from './PlotLineOrBand';
  *
  * */
 
-export interface PlotLineLabelOptions {
+export interface PlotBandLabelOptions {
     align?: AlignValue;
     formatter?: FormatUtilities.FormatterCallback<PlotLineOrBand>;
     rotation?: number;
@@ -43,17 +43,17 @@ export interface PlotLineLabelOptions {
     y?: number;
 }
 
-export interface PlotLineOptions {
+export interface PlotBandOptions {
     acrossPanes?: boolean;
+    borderColor?: ColorString;
+    borderWidth?: number;
     className?: string;
-    color?: ColorString;
-    dashStyle?: DashStyleValue;
+    color?: ColorType;
     events?: any;
+    from?: number;
     id?: string;
-    label?: PlotLineLabelOptions;
-    translatedValue?: number;
-    value?: number;
-    width?: number;
+    label?: PlotBandLabelOptions;
+    to?: number;
     zIndex?: number;
 }
 
@@ -63,4 +63,4 @@ export interface PlotLineOptions {
  *
  * */
 
-export default PlotLineOptions;
+export default PlotBandOptions;
