+++ ts/Extensions/CurrentDateIndication.ts
@@ -26,15 +26,15 @@ import type ColorString from '../Core/Color/ColorString';
 import type CSSObject from '../Core/Renderer/CSSObject';
 import type DashStyleValue from '../Core/Renderer/DashStyleValue';
 import type FormatUtilities from '../Core/FormatUtilities';
-import type { PlotBandLabelOptions } from '../Core/Axis/PlotBandOptions';
+import type { PlotBandLabelOptions } from '../Core/Axis/PlotLineOrBand/PlotBandOptions';
 import type {
     PlotLineLabelOptions,
     PlotLineOptions
-} from '../Core/Axis/PlotLineOptions';
+} from '../Core/Axis/PlotLineOrBand/PlotLineOptions';
 
 import Axis from '../Core/Axis/Axis.js';
 import Palette from '../Core/Color/Palette.js';
-import PlotLineOrBand from '../Core/Axis/PlotLineOrBand.js';
+import PlotLineOrBand from '../Core/Axis/PlotLineOrBand/PlotLineOrBand.js';
 import U from '../Core/Utilities.js';
 const {
     addEvent,
